+++ tests/test_load_optimizers.py
@@ -38,24 +38,3 @@ def test_load_optimizers_valid(valid_optimizer_names):
 def test_load_optimizers_invalid(invalid_optimizer_names):
     with pytest.raises(NotImplementedError):
         load_optimizers(invalid_optimizer_names)
-
-
-@pytest.mark.parametrize('optimizer_names', VALID_OPTIMIZER_NAMES)
-def test_learning_rate(optimizer_names):
-    with pytest.raises(ValueError):
-        optimizer = load_optimizers(optimizer_names)
-        optimizer(None, lr=-1e-2)
-
-
-@pytest.mark.parametrize('optimizer_names', VALID_OPTIMIZER_NAMES)
-def test_epsilon(optimizer_names):
-    with pytest.raises(ValueError):
-        optimizer = load_optimizers(optimizer_names)
-        optimizer(None, eps=-1e-6)
-
-
-@pytest.mark.parametrize('optimizer_names', VALID_OPTIMIZER_NAMES)
-def test_weight_decay(optimizer_names):
-    with pytest.raises(ValueError):
-        optimizer = load_optimizers(optimizer_names)
-        optimizer(None, weight_decay=-1e-3)
